factcheck-embeddings/src/factcheck_relevance/utils.py
import yaml
import json
import os

def load_config(config_path):
    with open(config_path, 'r') as f:
        return yaml.safe_load(f)

def load_json(file_path):
    with open(file_path, 'r') as f:
        return json.load(f)

def save_jsonl(data, file_path):
    os.makedirs(os.path.dirname(file_path), exist_ok=True)
    with open(file_path, 'w') as f:
        for item in data:
            f.write(json.dumps(item, ensure_ascii=False) + '\n')

def save_tsv(data, file_path):
    os.makedirs(os.path.dirname(file_path), exist_ok=True)
    with open(file_path, 'w') as f:
        for row in data:
            f.write('\t'.join(map(str, row)) + '\n')
